atv_demod: modify atv demod code [1/1]
PD#SWPL-2295

Problem:
1.modify atv demod code for coverity.
2.add ntsc-m ring filter for t962x2.
3.merge code optimization from android l.
4.add tuner compatible info in dts for tuner debug.

Solution:
1.modify atv demod code for coverity.
2.add ntsc-m ring filter for t962x2.
3.merge code optimization from android l.
4.add tuner compatible info in dts for tuner debug.

Verify:
verified by einstein


Conflicts:
	arch/arm/boot/dts/amlogic/txlx_t962x_r314.dts
	arch/arm64/boot/dts/amlogic/txlx_t962x_r314.dts
	drivers/amlogic/atv_demod/atv_demod_access.c
	drivers/amlogic/atv_demod/atv_demod_driver.c
	drivers/amlogic/atv_demod/atv_demod_ops.c
	drivers/amlogic/atv_demod/atvdemod_func.c

+++ drivers/amlogic/atv_demod/atv_demod_v4l2.c
@@ -587,7 +587,7 @@ static int v4l2_property_process_get(struct v4l2_frontend *v4l2_fe,
 	return 0;
 }
 
-static int v4l2_frontend_ioctl_properties(struct file *filp,
+static long v4l2_frontend_ioctl_properties(struct file *filp,
 			unsigned int cmd, void *parg)
 {
 	struct v4l2_frontend *v4l2_fe = video_get_drvdata(video_devdata(filp));
